Pull request feedback; renamed a bunch of stuff
