+++ sdk/core/azure-core-tracing-opentelemetry/cgmanifest.json
