+++ sdk/core/azure-core-tracing-opentelemetry/src/private/package_version.hpp
