+++ sdk/core/azure-core-tracing-opentelemetry/test/ut/azure_core_otel_test.cpp
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: MIT
 
 #define USE_MEMORY_EXPORTER 1
-#include "azure/core-opentelemetry/opentelemetry.hpp"
+#include "azure/core-tracing-opentelemetry/opentelemetry.hpp"
 #include <azure/core/test/test_base.hpp>
 
 #if defined(_MSC_VER)
