+++ sdk/core/azure-core-tracing-opentelemetry/vcpkg/Config.cmake.in
