+++ sdk/core/azure-core-tracing-opentelemetry/vcpkg/portfile.cmake
