+++ sdk/core/azure-core-tracing-opentelemetry/vcpkg/vcpkg.json
