+++ sdk/core/azure-core/src/context.cpp
@@ -12,7 +12,7 @@ Azure::DateTime Azure::Core::Context::GetDeadline() const
   // Contexts form a tree. Here, we walk from a node all the way back to the root in order to find
   // the earliest deadline value.
   auto result = DateTime::max();
-  for (std::shared_ptr<ContextSharedState> ptr = m_contextSharedState; ptr; ptr = ptr->Parent)
+  for (auto ptr = m_contextSharedState; ptr; ptr = ptr->Parent)
   {
     auto deadline = ContextSharedState::FromDateTimeRepresentation(ptr->Deadline);
     if (result > deadline)
